Finish variant types modulo any unknown bugs.
They can have type parameters and be recursive. There are 2
shift/reduce conflicts in the parser related to the 'destruct' form
(requiring 'end' at the end fixes it but I would rather avoid that..).

The datatype for type schemes has been removed and replaced by the
more general mechanism of type operators (type level abstractions and
application). Sometimes we still suggestively refer to a type as a
typescheme when it's being used in that way. E.g., we "instantiate" a
type abstraction acting as a type scheme by generating fresh type
variables and applying the abstraction to them. Similarly, we can
generalize a type by just adding binders for all of its free type
variables.

TODO: probably records next.